feat: compile cuda plugin statically
The patch modifies how plugin loading works by allowing plugins
to register themselves if their object file is statically linked
against the criu binary.

For now only the cuda plugin uses that since having it working
is the goal of this patch series but the system works for multiple
plugins.

If we contribute the patch upstream we have to think about adding
tests to it so that the criu test suite can validate the approach.